requirements.txt
pandas
requests
beautifulsoup4
streamlit
playwright
dnspython
python-dotenv
schedule
